requirements.txt
asgiref==3.5.1
Django==4.0.4
djangorestframework==3.13.1
lxml==4.9.2
mysqlclient==2.1.1
PyMySQL==1.0.2
pytz==2022.1
spyne==2.14.0
sqlparse==0.4.2
suds==1.1.2
tzdata==2022.1
virtualenv==20.21.0
